# Speed up correct_k