Profiling the chemdose_ph_chain examples from the vignettes, `correct_k` was noticed to be slow:
![image](https://github.com/user-attachments/assets/bd027c0d-b0cf-4813-a40d-b7c2625dcf75)

Applying optimizations contained within:

![image](https://github.com/user-attachments/assets/7f2ac833-2dfd-437a-b590-7c229123c396)


`correct_k` goes from 1980ms to 90ms (-95.4% faster)

Overall example goes from 7.74s to 5.67s (-27% faster).

These changes and gains are independent of: https://github.com/BrownandCaldwell-Public/tidywater/pull/7