@Carter12s 2024-11-10
This this MR with #7 yields:

![image](https://github.com/user-attachments/assets/7a0aa3fe-98f4-46c3-a7be-f580883d30cf)

7.7s -> 2.2s  = -68% faster